[20497] Changed agency ID type to integer instead of long.
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@103255 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ harvester/rr/src/main/java/dk/dbc/dataio/harvester/rr/AgencyConnection.java
@@ -41,7 +41,7 @@ public class AgencyConnection {
     private static final Logger LOGGER = LoggerFactory.getLogger(AgencyConnection.class);
 
     private final OpenAgencyConnector connector;
-    private final HashMap<Long, LibraryRules> libraryRulesCache;    // unbounded in-memory cache
+    private final HashMap<Integer, LibraryRules> libraryRulesCache;    // unbounded in-memory cache
 
     /**
      * Class constructor
@@ -70,7 +70,7 @@ public class AgencyConnection {
      * @return library rules or null if no rules could be found.
      * @throws IllegalStateException on error communicating with the OpenAgency web-service
      */
-    public LibraryRules getLibraryRules(long agencyId, String trackingId) throws IllegalStateException {
+    public LibraryRules getLibraryRules(int agencyId, String trackingId) throws IllegalStateException {
         return libraryRulesCache.computeIfAbsent(agencyId, id -> {
             try {
                 final Optional<dk.dbc.oss.ns.openagency.LibraryRules> response = connector.getLibraryRules(id, trackingId);
